refactor: Export pages directly from index file
